Got hello-world working, then probably messed it up.
Got PIZZAHUT to output valid XML files with the dummy data it was given,
then made a bunch of changes to increase what it outputs. Need a real
test replay.

During this work it became clear that it's unlikely possible to acquire
most of the statistics I wanted to collect (resources collected, workers
built, average income, etc.). It appears those are stored by the game
client, but not in the replay proper. I could try to derive it, but I
wouldn't bet on it, especially for a 1.0 deployment.

But now, to bed.