Embedded images rather than loaded at runtime.
While it inflates the SWF size, a) it's much, much easier (the Loaders
were completely opaque and unreliable, and required a ton of code to
maintain), and b) makes the SWF more portable.  If anyone else wants
to use this, at the very least things like Race and League icons can be
embedded into the SWF without having to fiddle with things like
paths, especially since we have no idea how this will resolve those kinds
of issues when put into XSplit.