PizzaHut build break, more restructuring.
Actionscript is a shitshow. I've modeled this after 3 different
tutorials and the box model is nigh impossible. Going to try
simply removing TacoRenderer as a separate part of the Displaylist
and just attaching all the graphical components onto the main Sprite,
since most of the blank screens I'm getting seem to be trying to
render the Renderer at 0 width, 0 height and resizing at various points
is bunk.