Removed TacoRenderer, seeing some progress.
In a flare of fury, I removed the entire TacoRenderer class thinking
that every layer or DisplayList abstraction is just confusing the hell
out of me. The table is now getting drawn _more or less_ like it should
be, with a few exceptions:

* For some reason, (x,y) for both the main sprite and the PlayerTable
start near the center of the Stage.
* A call to drawRect to add a background color to an element was fucking
up the width. I've kept the trace msgs, but that's the moment when it
explodes and thought it's width was 25,000. Apparently, drawRect is
side-effectful.

Still a few other bugs: the game parser seems to believe there is only
one player in the test replay, and it's still ugly as sin. But once I
can render comfortably this should be golden.